azure_function/requirements.txt
azure-functions>=1.17
azure-identity>=1.14
azure-keyvault-secrets>=4.7
cryptography>=41.0
